pyproject.toml
[project]
name = "06-nerf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pydantic>=2.12.5",
    "pytorch-lightning>=2.6.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "wandb>=0.25.0",
]
